digi.js: deduplicated +1/+2 portrait images for recent Digimon
For older Digimon, the game files (from which digi.js is now generated)
already use the same images for +1 and +2 portaits, but for recently
added Digimon for some reason they use distinct (but identical) images.
To avoid downloading identical images twice, replace the +2 image IDs
with the +1 IDs.

This matches the current structure of img/mon, which has no +2 images.