better, and numerically indexed, monster portrait images
The new images are:

1. extracted from the files the game client downloads from the game
   server upon initial install and upon the start of each new event,

2. converted to indexed color using modern pngquant at quality
   level 45 (selected to keep the total size of the new images
   from exceeding the total size of the old images),

3. further slimmed down losslessly by a combination of pingo,
   Efficient Compression Tool, DeflOpt and defluff.

Individual portraits are stored separately from the common backgrounds
and frames that are shared by all Digimon at each evolution stage.
Each Digimon's full portrait will thus be composed of three layers
in the front-end: background, portrait proper, frame.

The "empty" and "none" placeholders are similar but have only two layers:
background (which already displays the word "empty" or "none") and frame.

Frames and backgrounds are compressed less heavily: they are reduced
to 256 colors unless already indexed or grayscale, and backing/5.png is not
reduced at all because the quality loss is too great even at 256 colors.
They all go through the lossless optimizers just as the portrait images do.

As a result, the new images look much better than the old ones.

The new file names correspond to the values in the `images` array property
of each monster object in the new digi.js.